feat: 추가 구매 또는 종료 구현
