+++ README.md
@@ -134,15 +134,15 @@
 
 ### 🛒 추가 구매 또는 종료
 
-- [ ] 추가 구매 여부를 확인하기 위해 안내 문구를 출력한다
-- [ ] 추가 구매 여부를 입력 받는다
-  - [ ] Y
-    - [ ] 재고가 업데이트된 상품 목록을 확인 후 추가로 구매를 진행한다
-  - [ ] N
-    - [ ] 구매를 종료한다
-  - [ ] 예외 처리
-    - [ ] 입력이 `Y` 또는 `N`이 아닐 경우
-  - [ ] 사용자가 잘못된 값을 입력했을 때 `IllegalArgumentException`을 발생시키고
+- [x] 추가 구매 여부를 확인하기 위해 안내 문구를 출력한다
+- [x] 추가 구매 여부를 입력 받는다
+  - [x] Y
+    - [x] 재고가 업데이트된 상품 목록을 확인 후 추가로 구매를 진행한다
+  - [x] N
+    - [x] 구매를 종료한다
+  - [x] 예외 처리
+    - [x] 입력이 `Y` 또는 `N`이 아닐 경우
+  - [x] 사용자가 잘못된 값을 입력했을 때 `IllegalArgumentException`을 발생시키고
     <br>`[ERROR]`로 시작하는 오류 메시지를 출력 후 다시 입력 받는다
