+++ src/main/java/store/util/Constant.java
@@ -1,0 +1,19 @@
+package store.util;
+
+public class Constant {
+
+    public static final String PRODUCTS_FILE_NAME = "src/main/resources/products.md";
+    public static final String PROMOTIONS_FILE_NAME = "src/main/resources/promotions.md";
+    public static final String BLANK = "";
+
+    public static final String WELCOME_MESSAGE = "안녕하세요. W편의점입니다.";
+    public static final String INTRODUCE_PRODUCT_MESSAGE = "현재 보유하고 있는 상품입니다.";
+    public static final String ORDER_MESSAGE = "구매하실 상품명과 수량을 입력해 주세요. (예: [사이다-2],[감자칩-1])";
+    public static final String MEMBERSHIP_DISCOUNT_MESSAGE = "멤버십 할인을 받으시겠습니까? (Y/N)";
+    public static final String ADDITIONAL_PURCHASE_MESSAGE = "감사합니다. 구매하고 싶은 다른 상품이 있나요? (Y/N)";
+
+    public static final String RECEIPT_TITLE = "==============W 편의점================";
+    public static final String RECEIPT_PRODUCT_NAME_QUANTITY_AMOUNT = "상품명                수량      금액";
+    public static final String RECEIPT_GIVEAWAY_TITLE = "=============증\t\t정===============";
+    public static final String RECEIPT_AMOUNT_INFO_TITLE = "====================================";
+}
