--- src/main/java/store/view/Constant.java
@@ -1,12 +1,0 @@
-package store.view;
-
-public class Constant {
-
-    public static final String PRODUCTS_FILE_NAME = "src/main/resources/products.md";
-    public static final String PROMOTIONS_FILE_NAME = "src/main/resources/promotions.md";
-    public static final String BLANK = "";
-
-    public static final String WELCOME_MESSAGE = "안녕하세요. W편의점입니다.";
-    public static final String INTRODUCE_PRODUCT_MESSAGE = "현재 보유하고 있는 상품입니다.";
-    public static final String ORDER_MESSAGE = "구매하실 상품명과 수량을 입력해 주세요. (예: [사이다-2],[감자칩-1])";
-}
